Update code.py
For some reason, I could not get the Adafruit Mini GPS to find a signal while other I2C sensors are connected. A few things were suggested. such as; injecting power to the GPS, adding a filter capacitor, and having multiple copies of a line of code ("gps.update()"). I've tried every combination of those suggestions and none of them allow me to pick up a signal. 

If the Mini GPS is connected directly to the Pynt with no other sensors connected, it works fine - picks up a signal within 30 seconds. But as soon as I add a sensor, the fix quality remains at zero. If you have any other suggestions, please, please let us know. 

On this fork, you do need to download the "backgraph-planet.bmp" file in the Images folder or remove or replace it with your own. Have fun and happy Treking!